新建客户字段非空限制

+++ modules/118_ccc_ccc/lib/ccc/incident_journals_controller_ex.rb
@@ -98,10 +98,16 @@ module Ccc::IncidentJournalsControllerEx
               # 此处评论创建成功
               if !@incident_request.incident_status_id.eql?("000K000A0g8zPKXoIwOIhk") && !@incident_request.incident_status_id.eql?("000K000A0g9LO0pOKPsZ1s")
                 if sla_instance_id.present?
-                  sla_instance = Slm::SlaInstance.find(sla_instance_id)
-                  sa = Slm::ServiceAgreement.find(sla_instance.service_agreement_id)
-                  Slm::SlaInstance.start(sa,{:bo_type => "Icm::IncidentRequest", :bo_id => @incident_request.id, :service_agreement_id => sa.id})
-                  sla_instance.destroy
+                  sla_instance = Slm::SlaInstance.where("id = ?",sla_instance_id)
+                  if sla_instance.first().present?
+                    sla_instance = sla_instance.first()
+                    sa = Slm::ServiceAgreement.where("id = ?",sla_instance)
+                    if sa.first().present?
+                      sa = sa.first()
+                      Slm::SlaInstance.start(sa,{:bo_type => "Icm::IncidentRequest", :bo_id => @incident_request.id, :service_agreement_id => sa.id})
+                      sla_instance.destroy
+                    end
+                  end
                 end
               end
             end
